Convert latitudes and longitudes into radians
As the title describes this commit comprises of converting the latitudes and longitudes into radians to be able to use in the clustering algorithm - DBSCAN. Why? The idea is to group the nearby latitude and longitudes together based on some predefined distance metric. DBSCAN does not need any predefined number of clusters as it is dynamic in figuring out the clusters starting from denser side of data points.